+++ node/clients/index.ts
@@ -1,16 +1,15 @@
-import { IOClients } from '@vtex/api'
+import { IOClients } from "@vtex/api";
+import { Checkout } from "@vtex/clients";
 
-import { CustomPrice } from './customPrice'
-import Session from './session'
+import { CustomPrice } from "./customPrice";
 
 // Extend the default IOClients implementation with our own custom clients.
 export class Clients extends IOClients {
-
   public get customPrice() {
-    return this.getOrSet('customPrice', CustomPrice)
+    return this.getOrSet("customPrice", CustomPrice);
   }
 
-  public get customSession() {
-    return this.getOrSet('customSession', Session)
+  public get checkout() {
+    return this.getOrSet("checkout", Checkout);
   }
 }