+++ node/resolvers/orderConfiguration/enums/DataEntity.ts
@@ -1,0 +1,4 @@
+export enum DataEntity {
+    CLIENT = "CL",
+    ORDER_CONFIG = "order_configuration",
+}