+++ node/resolvers/orderConfiguration/index.ts
@@ -1,0 +1,14 @@
+import { saveOrderConfiguration } from "./saveOrderConfiguration";
+import {orderConfiguration} from "./orderConfiguration";
+
+export const mutations = {
+  saveOrderConfiguration
+};
+
+export const queries = {
+  /**
+   * Get user session custom keys
+   * @return String
+   */
+  orderConfiguration
+};