+++ node/utils/vtexId.ts
@@ -1,13 +1,18 @@
-const vtexIdAccount = (account: string) => `VtexIdclientAutCookie_${account}`
-const VTEX_ID = 'VtexIdclientAutCookie'
+const vtexIdAccount = (account: string) => `VtexIdclientAutCookie_${account}`;
+const VTEX_ID = "VtexIdclientAutCookie";
 
 export const vtexIdCookies = (ctx: Context) => {
-  const { cookies, vtex: { account }} = ctx
-  const vtexIdAccountCookieValue = cookies.get(vtexIdAccount(account))
-  const vtexIdValue = cookies.get(VTEX_ID)
+  const {
+    cookies,
+    vtex: { account }
+  } = ctx;
+  const vtexIdAccountCookieValue = cookies.get(vtexIdAccount(account));
+  const vtexIdValue = cookies.get(VTEX_ID);
 
   return {
-    account: vtexIdAccountCookieValue ? `${vtexIdAccount(account)}=${vtexIdAccountCookieValue}` : null,
-    id: vtexIdValue ? `${VTEX_ID}=${vtexIdValue}` : null,
-  }
-}
+    account: vtexIdAccountCookieValue
+      ? `${vtexIdAccount(account)}=${vtexIdAccountCookieValue}`
+      : null,
+    id: vtexIdValue ? `${VTEX_ID}=${vtexIdValue}` : null
+  };
+};