+++ react/FormInputText.tsx
@@ -6,13 +6,11 @@ import { FormRawInputProps, InputTypes } from './typings/InputProps'
 export default function FormInput(props: FormRawInputProps) {
   const { inputType = InputTypes.input, ...rest } = props
 
-  // eslint-disable-next-line default-case
-  switch (inputType) {
-    case InputTypes.input:
-      return <RawInput {...rest} pointer={props.pointer} />
-    case InputTypes.hidden:
-      return <HiddenInput {...rest} pointer={props.pointer} />
-    case InputTypes.password:
-      return <PasswordInput {...rest} pointer={props.pointer} />
+  const INPUT_TYPES = {
+    [InputTypes.input]: <RawInput {...rest} pointer={props.pointer} />,
+    [InputTypes.hidden]: <HiddenInput {...rest} pointer={props.pointer} />,
+    [InputTypes.password]: <PasswordInput {...rest} pointer={props.pointer} />,
   }
+
+  return INPUT_TYPES[inputType]
 }